[DependencyInjection] Move MultiParentingToAbstractDependencyRector here
