+++ composer.json
@@ -80,6 +80,7 @@
             "Rector\\ConsoleDiffer\\": "packages/console-differ/src",
             "Rector\\Core\\": "src",
             "Rector\\DeadCode\\": "rules/dead-code/src",
+            "Rector\\DependencyInjection\\": "rules/dependency-injection/src",
             "Rector\\EarlyReturn\\": "rules/early-return/src",
             "Rector\\DeadDocBlock\\": "rules/dead-doc-block/src",
             "Rector\\DoctrineAnnotationGenerated\\": "packages/doctrine-annotation-generated/src",
@@ -207,6 +208,7 @@
             "Rector\\Core\\Tests\\": "tests",
             "Rector\\EarlyReturn\\Tests\\": "rules/early-return/tests",
             "Rector\\DeadCode\\Tests\\": "rules/dead-code/tests",
+            "Rector\\DependencyInjection\\Tests\\": "rules/dependency-injection/tests",
             "Rector\\DoctrineCodeQuality\\Tests\\": "rules/doctrine-code-quality/tests",
             "Rector\\DoctrineGedmoToKnplabs\\Tests\\": "rules/doctrine-gedmo-to-knplabs/tests",
             "Rector\\Doctrine\\Tests\\": "rules/doctrine/tests",
