+++ rules/dependency-injection/tests/Rector/Class_/MultiParentingToAbstractDependencyRector/Source/AnotherDependency.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\DependencyInjection\Tests\Rector\Class_\MultiParentingToAbstractDependencyRector\Source;
+
+final class AnotherDependency
+{
+
+}
