+++ rules/dependency-injection/tests/Rector/Class_/MultiParentingToAbstractDependencyRector/Source/SecondImplementer.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\DependencyInjection\Tests\Rector\Class_\MultiParentingToAbstractDependencyRector\Source;
+
+class SecondImplementer implements ImplementInterface
+{
+
+}
