--- rules/solid/tests/Rector/Class_/MultiParentingToAbstractDependencyRector/Source/ImplementInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SOLID\Tests\Rector\Class_\MultiParentingToAbstractDependencyRector\Source;
-
-interface ImplementInterface
-{
-
-}
