--- rules/solid/tests/Rector/Class_/MultiParentingToAbstractDependencyRector/Source/SecondImplementer.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SOLID\Tests\Rector\Class_\MultiParentingToAbstractDependencyRector\Source;
-
-class SecondImplementer implements ImplementInterface
-{
-
-}
